Cleanup some declarations & automate delay(3)

+++ kos/src/libc/user/kos.except-handler.c
@@ -28,7 +28,7 @@ DECL_BEGIN
 
 /*[[[start:implementation]]]*/
 
-/*[[[head:libc_set_exception_handler,hash:CRC-32=0x2048d50a]]]*/
+/*[[[head:libc_set_exception_handler,hash:CRC-32=0x7ca64b8b]]]*/
 /* Set the exception handler mode for the calling thread.
  * Examples:
  *     Set mode #1: set_exception_handler(EXCEPT_HANDLER_MODE_DISABLED, NULL, NULL)
@@ -45,7 +45,7 @@ DECL_BEGIN
  * @return: 0 :        Success.
  * @return: -1:EINVAL: The given MODE is invalid */
 INTERN int
-NOTHROW(LIBCCALL libc_set_exception_handler)(int mode,
+NOTHROW(LIBCCALL libc_set_exception_handler)(unsigned int mode,
                                              except_handler_t handler,
                                              void *handler_sp)
 /*[[[body:libc_set_exception_handler]]]*/
@@ -58,7 +58,7 @@ NOTHROW(LIBCCALL libc_set_exception_handler)(int mode,
 }
 /*[[[end:libc_set_exception_handler]]]*/
 
-/*[[[head:libc_get_exception_handler,hash:CRC-32=0x83658349]]]*/
+/*[[[head:libc_get_exception_handler,hash:CRC-32=0x36c47e2f]]]*/
 /* Get the current exception handler mode for the calling thread.
  * @param: PMODE:       When non-NULL, store the current mode, which is encoded as:
  *                       - One of `EXCEPT_HANDLER_MODE_(DISABLED|ENABLED|SIGHAND)'
@@ -73,7 +73,7 @@ NOTHROW(LIBCCALL libc_set_exception_handler)(int mode,
  * @return: 0 :         Success.
  * @return: -1:EFAULT:  One of the given pointers is non-NULL and faulty */
 INTERN int
-NOTHROW_NCX(LIBCCALL libc_get_exception_handler)(int *pmode,
+NOTHROW_NCX(LIBCCALL libc_get_exception_handler)(unsigned int *pmode,
                                                  except_handler_t *phandler,
                                                  void **phandler_sp)
 /*[[[body:libc_get_exception_handler]]]*/
@@ -89,7 +89,7 @@ NOTHROW_NCX(LIBCCALL libc_get_exception_handler)(int *pmode,
 	                                   phandler,
 	                                   phandler_sp);
 	if (likely(E_ISOK(result)) && pmode)
-		*pmode = mode_value;
+		*pmode = (unsigned int)mode_value;
 #endif /* __SIZEOF_SYSCALL_LONG_T__ != __SIZEOF_INT__ */
 	return libc_seterrno_syserr(result);
 }
